+++ src/ScrapperImplemantations/SkinPort_Scrapper.cs
@@ -12,7 +12,29 @@ namespace CSSkinScrapper.ScrapperImplemantations
     {
         protected override string baseUrl => "https://api.skinport.com/v1/items?app_id=730&currency=EUR&tradable=0";
 
-        private string completeMarket;
+        //TODO: share this between this and SkinPortImpl. -> start one async method at start and check if its completed if result is needed
+        protected string completeMarket
+        {
+            get
+            {
+                if (m_completeMarket == "")
+                {
+                    using (var client = new HttpClient())
+                    {
+                        var response = client.GetAsync("https://api.skinport.com/v1/items?app_id=730&currency=EUR&tradable=0").GetAwaiter().GetResult();
+                        if (!response.IsSuccessStatusCode)
+                        {
+                            //TODO: dont throw up
+                            throw new Exception("Couldn't get the market from skinport.");
+                        }
+                        m_completeMarket = response.Content.ReadAsStringAsync().GetAwaiter().GetResult();
+                    }
+                }
+                return m_completeMarket;
+            }
+        }
+
+        private string m_completeMarket = "";
 
         internal SkinPort_Scrapper()
         {
@@ -23,7 +45,7 @@ namespace CSSkinScrapper.ScrapperImplemantations
                 throw new Exception();
             }
 
-            completeMarket = response.Content.ReadAsStringAsync().GetAwaiter().GetResult();
+            m_completeMarket = response.Content.ReadAsStringAsync().GetAwaiter().GetResult();
         }
 
         public override double GetPrice(string skin)
