+++ src/SkinScrapper.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System;
 using CSSkinScrapper.Interop;
+using System.Net.Http;
 
 namespace CSSkinScrapper
 {
@@ -18,19 +19,34 @@ namespace CSSkinScrapper
 
         public static double[] GetPriceArray(List<Skin> skins)
         {
-            return instance.scrapperList[0].GetPriceArray(skins);
+            return instance.scrapperList[0].GetPriceArray(skins).Result;
 
             instance.GetPriceArrays(skins);
         }
 
-        private double[] GetPriceArrays(List<Skin> skins)
+        private static HttpClient? userClient;
+        public static bool SkinExists(Skin skin)
+        {
+            if (userClient is null)
+                userClient = new HttpClient();
+
+            string req = instance.scrapperList[0].GetUrl(skin);
+            var resp = userClient.GetAsync(req).Result;
+
+            if (!resp.IsSuccessStatusCode)
+                return false;
+            else
+                return true;
+        }
+
+        private (double[] steam, double[]skinport) GetPriceArrays(List<Skin> skins)
         {
             foreach (var scrapper in scrapperList)
             {
                 
             }
 
-            return null;
+            return (null, null);
         }
     }
 }
